requirements.txt
numpy
pandas
matplotlib
scipy
torch
batman-package
emcee
corner
tqdm
jupyter
scikit-learn